refactor: use DomPDF to save PDF (WFS-715)
The PDF will be generated by DomPDF. The data is sent via the URL, therefore our ODF Processor has access to the data, which comes from the form.
In case of the event list, the data will be fetched from the EventRepository.

For Contactforms and Eventforms, all the data is read from the url.