feat: extend condition for showing publish action
Check if direct publishing is allowed and if the current record
is a workspace version (t3ver_wsid > 0) and if the workspace stage is either
"Editing" (t3ver_stage === 0) or any custom stage (t3ver_stage > 0).
Remove the publish button from ReadyToPublish partial
as this is not the right place for it.